chore: create setupBasicSeries function to reuse throughout chart widgets unit tests
